tire identify/requirements.txt
streamlit
pillow
numpy
scikit-image
